Large scale refactor to existing tests + Yaml files to create pipelines (#42)
* refactor(Full-Project): Initial file reorganisation

Moved and adjusted all imported settings to allow compilation. Project should work as it did before reorganisation

* refactor(generic-keywords): Split all generic keywords into specific files with a dedicated function

This will be a breaking change if we remove the generic-keywords.robot file.
None of the new files have been refactored to work with test as of yet.
This is simply just the "skeleton" of what the structure will be.

* refactor(keywords): Fully phased out Generic-keyword.robot

Fully split away from generic-keywords.robot with all suites still passing.
Next in order is to format the keywords themselves and move to more fitting locations.

* refactor(Keywords): In the middle of big scale refactor of keywords

Moving low level func and page specific func to their own file. Likewise with navigation

BREAKING CHANGE: Some of the Scenarios are currently broken due to refactoring

* refactor(Keywords-POM): Snåla kalle and kim family scenarios are fully functional with the refactoring

* refactor(Snåla-Kalle): Removed reused code and keywords and baked it into the base keywords

* refactor(Snåla-kalle): Removed 1 redundant keyword.

* Fully phased out Kim-Keywords. (#39)

* Merge pull request #40

* refactor(Snåla-Kalle): Cleaning comments, removed embedded arguments …

* refactor(Snåla-Kalle): Moved price checking function to cart-keywords…

* refactor(Pathing): Added EXECDIR to all file pathings instead of relative

* docs(PDF-Doc): Moved assignment files into their own directory

- moved project 1 files into asset dir
- added project 2 files into asset dir

* feat(Yaml): Added yaml files for pipelines

This is the "raw" version of the yaml files that we got for the project

* style(Folder-Names): Renamed some folders to be more correct to standard conventions

* feat: Added "User Registers" test as a new-feature

This is just to see if our pipline is working (also only tagging 1 to avoid too much usage on Azure)

* fix(yaml-fix): fixed naming for test folder so that azure can access the target folder.

* refactor(Tags): Added new-feature tags to all current tests

---------

Co-authored-by: berggren00 <145928987+berggren00@users.noreply.github.com>
Co-authored-by: Daniel Kvist <61018006+Slomokv@users.noreply.github.com>